Modified script to include m365 endpoint versions, added additional .pac files
The script now contains a mandatory parameter for the tenantname. Also writing a header containing the version of the m365 endpoints web api version so we can keep track of that and check if newer endpoints should be expected. 

Generated additional pac files, now containing the tenant name and the used instance in the pac filename. Update used config in Intune to use the new pac file(s).